requirements.txt
python-dotenv==1.0.1
requests==2.32.3
obs-websocket-py==1.0
websocket-client==1.9.0
